+++ findNumbers.js
@@ -1,7 +1,0 @@
-'use strict'
-
-// Not everyone needs `findNumbers()` function,
-// so to reduce application bundle size
-// it has been moved to a separate file.
-
-module.exports = require('./build/findNumbers')