+++ source/findNumbers/Leniency.js
@@ -1,5 +1,6 @@
 import parseNumber from '../parse'
 import isValidNumber from '../validate'
+import { parseDigit } from '../common'
 
 import
 {
@@ -19,7 +20,7 @@ export default
 	 */
 	POSSIBLE(number, candidate, metadata)
 	{
-		return parseNumber(number, { extended: true }, metadata).possible
+		return true
 	},
 
 	/**
@@ -35,7 +36,9 @@ export default
 			return false
 		}
 
-		return isNationalPrefixPresentIfRequired(number, metadata)
+    // Skipped for simplicity.
+		// return isNationalPrefixPresentIfRequired(number, metadata)
+    return true
   },
 
 	/**
@@ -131,7 +134,7 @@ function containsOnlyValidXChars(number, candidate, metadata)
 	      // This is the extension sign case, in which the 'x' or 'X' should always precede the
 	      // extension number.
       }
-      else if (!PhoneNumberUtil.normalizeDigitsOnly(candidate.substring(index)) === number.getExtension())
+      else if (parseDigits(candidate.substring(index)) !== number.ext)
       {
         return false
       }
@@ -162,8 +165,8 @@ function isNationalPrefixPresentIfRequired(number, _metadata)
   const nationalNumber = util.getNationalSignificantNumber(number)
   const formatRule = util.chooseFormattingPatternForNumber(metadata.numberFormats(), nationalNumber)
 
-  // To do this, we check that a national prefix formatting rule was present and that it wasn't
-  // just the first-group symbol ($1) with punctuation.
+  // To do this, we check that a national prefix formatting rule was present
+  // and that it wasn't just the first-group symbol ($1) with punctuation.
   if (formatRule && formatRule.getNationalPrefixFormattingRule().length > 0)
   {
     if (formatRule.getNationalPrefixOptionalWhenFormatting())
@@ -391,4 +394,25 @@ function allNumberGroupsRemainGrouped
   // match the last group of the subscriber number. Note the extension cannot have
   // formatting in-between digits.
   return normalizedCandidate.slice(fromIndex).contains(number.getExtension())
+}
+
+function parseDigits(string)
+{
+  let result = ''
+
+  // Using `.split('')` here instead of normal `for ... of`
+  // because the importing application doesn't neccessarily include an ES6 polyfill.
+  // The `.split('')` approach discards "exotic" UTF-8 characters
+  // (the ones consisting of four bytes) but digits
+  // (including non-European ones) don't fall into that range
+  // so such "exotic" characters would be discarded anyway.
+  for (const character of string.split(''))
+  {
+    const digit = parseDigit(character)
+    if (digit) {
+      result += digit
+    }
+  }
+
+  return result
 }