+++ source/findPhoneNumbers.js
@@ -13,7 +13,7 @@ from './common'
 
 import parsePreCandidate from './findNumbers/parsePreCandidate'
 import isValidPreCandidate from './findNumbers/isValidPreCandidate'
-import isValidCandidate, { OPENING_PARENS } from './findNumbers/isValidCandidate'
+import isValidCandidate from './findNumbers/isValidCandidate'
 
 // Copy-pasted from `./parse.js`.
 const VALID_PHONE_NUMBER =
@@ -30,7 +30,7 @@ const VALID_PHONE_NUMBER =
 const EXTN_PATTERNS_FOR_PARSING = create_extension_pattern('parsing')
 
 const WHITESPACE_IN_THE_BEGINNING_PATTERN = new RegExp('^[' + WHITESPACE + ']+')
-const WHITESPACE_AND_OPENING_PARENS_IN_THE_END_PATTERN = new RegExp('[' + WHITESPACE + OPENING_PARENS + ']+$')
+const PUNCTUATION_IN_THE_END_PATTERN = new RegExp('[' + VALID_PUNCTUATION + ']+$')
 
 // // Regular expression for getting opening brackets for a valid number
 // // found using `PHONE_NUMBER_START_PATTERN` for prepending those brackets to the number.
@@ -125,7 +125,7 @@ export class PhoneNumberSearch
 		// Fixes not parsing numbers with whitespace in the end.
 		// Also fixes not parsing numbers with opening parentheses in the end.
 		// https://github.com/catamphetamine/libphonenumber-js/issues/252
-		number = number.replace(WHITESPACE_AND_OPENING_PARENS_IN_THE_END_PATTERN, '')
+		number = number.replace(PUNCTUATION_IN_THE_END_PATTERN, '')
 
 		number = parsePreCandidate(number)